File into main and renderer
