--- src/actions/SearchWordActions.ts
@@ -1,9 +1,0 @@
-import { actionCreatorFactory } from 'typescript-fsa';
-
-const actionCreator = actionCreatorFactory('search-word-action');
-
-export const changeSearchWordAction = actionCreator<Partial<string>>(
-  'change-search-word',
-);
-
-export default changeSearchWordAction;
