+++ src/renderer/actions/SelectedWordsActions.ts
