+++ src/renderer/components/ActivityBar.tsx
