+++ src/renderer/components/BookViewContainer.tsx
