+++ src/renderer/components/ContentEditable.tsx
