+++ src/renderer/components/ControlBox.tsx
