+++ src/renderer/components/Editor.tsx
