+++ src/renderer/components/PrimarySidebar.tsx
