+++ src/renderer/components/SecondarySidebar.tsx
