+++ src/renderer/components/StatusBar.tsx
