+++ src/renderer/components/WordCard.tsx
