+++ src/renderer/components/WordTabs.tsx
