+++ src/renderer/reducers/PrimarySidebarReducer.ts
