+++ src/renderer/reducers/SecondarySidebarReducer.ts
