+++ src/renderer/reducers/SelectedWordsReducer.ts
