+++ src/renderer/states/SelectedWord.ts
