+++ tsconfig.json
@@ -68,6 +68,6 @@
     "forceConsistentCasingInFileNames": true  /* Disallow inconsistently-cased references to the same file. */,
   },
   "files": [
-    "src/main.ts"
+    "src/main/index.ts"
   ]
 }
